Move hip specific header files to include/hip and use USE_ROCM instead of HIP_PLATFORM_AMD
