+++ sgl-kernel/include/hip/impl/hip_vec_fp32_impl.h
