+++ sgl-kernel/include/hip/impl/hip_vec_half_impl.h
