+++ sgl-kernel/setup_rocm.py
@@ -80,7 +80,6 @@ hipcc_flags = [
     "-Xcompiler",
     "-fPIC",
     "-std=c++17",
-    "-D__HIP_PLATFORM_AMD__=1",
     f"--amdgpu-target={amdgpu_target}",
     "-DENABLE_BF16",
     "-DENABLE_FP8",
